Store nodes in a matrix instead of a list
This improves the performance of the solver since array accessing is faster
than indexing into a list.

Change the Heap to use a derived class for specifying the comparer instead of
using a delegate. This reduces the number of allocations and also speeds up
the solver.

Use direct bitwise checking in the grid for the possible values of a node.
This reduces allocations since previously an interator was allocated for each
node everytime we were deciding which possible values it has.