Improve entropy for generating Sudoku puzzles
The previous version was randomizing the creation of the puzzle by randomizing the order in which values are checked as possible values for a node.
However, that randomization was only happening once, at the beginning of the puzzle creation and was then re-used for each subsequent selection of a value.

With this change the order of picking values is changed for each new node. This introduces a small perf hit as values need to be re-computed every single time, however the increase in
the randomized puzzles is worth it.